+++ src/app/elements/lookup/simple/lookup.component.ts
@@ -1,12 +1,12 @@
 import { Component } from '@angular/core';
-import { BaseLookupComponent } from './baseLookup';
+import { BaseLookupComponent } from '../baseLookup';
 
 @Component({
   /* tslint:disable:component-selector */
   selector: 'lookup',
   /* tslint:enable:component-selector*/
   templateUrl: './lookup.component.html',
-  providers: [
-  ]
+  styleUrls: ['../baseLookup.css'],
+  providers: []
 })
 export class LookupComponent extends BaseLookupComponent { }