+++ src/app/elements/charts/finalResults/finalResults.charts.component.ts
@@ -1,19 +1,24 @@
-import { Component, Input, ViewChild, OnInit, OnChanges, Output, EventEmitter } from '@angular/core';
+import {
+  Component,
+  Input,
+  ViewChild,
+  OnInit,
+  OnChanges,
+  Output,
+  EventEmitter
+} from '@angular/core';
 import { SimpleRequester } from '../../../services/simple-requester';
 import { TestResultService } from '../../../services/test-result.service';
 import { TestResult } from '../../../shared/models/test-result';
 import { FinalResult } from '../../../shared/models/final-result';
 import { FinalResultService } from '../../../services/final_results.service';
 import { BaseChartDirective } from 'ng2-charts';
+import { colors } from '../../../shared/colors.service';
 
 @Component({
   selector: 'final-result-chart',
   templateUrl: './finalResults.charts.component.html',
-  providers: [
-    SimpleRequester,
-    TestResultService,
-    FinalResultService
-  ]
+  providers: [SimpleRequester, TestResultService, FinalResultService]
 })
 export class FinalResultChartsComponent implements OnInit, OnChanges {
   @ViewChild(BaseChartDirective) chart: BaseChartDirective;
@@ -35,9 +40,7 @@ export class FinalResultChartsComponent implements OnInit, OnChanges {
     }
   };
 
-  constructor(
-    private finalResultService: FinalResultService,
-  ) { }
+  constructor(private finalResultService: FinalResultService) {}
 
   async ngOnInit() {
     this.listOfFinalResults = await this.finalResultService.getFinalResult({});
@@ -66,52 +69,63 @@ export class FinalResultChartsComponent implements OnInit, OnChanges {
   fillChartLabels() {
     this.doughnutChartLabels = [];
     for (const finalResult of this.listOfFinalResults) {
-      this.doughnutChartLabels.push(finalResult.name + ' | ' + this.calculatePrecentageAndCount(finalResult.name));
+      this.doughnutChartLabels.push(
+        finalResult.name +
+          ' | ' +
+          this.calculatePrecentageAndCount(finalResult.name)
+      );
     }
   }
 
   fillChartData() {
     this.doughnutChartData = [];
     for (const finalResult of this.listOfFinalResults) {
-      this.doughnutChartData.push(this.testResults.filter(x => x.final_result.name === finalResult.name).length);
+      this.doughnutChartData.push(
+        this.testResults.filter(x => x.final_result.name === finalResult.name)
+          .length
+      );
     }
   }
 
   calculatePrecentageAndCount(finalResult: String): String {
-    const num = this.testResults.filter(x => x.final_result.name === finalResult).length;
-    const percentage = num / this.testResults.length * 100;
+    const num = this.testResults.filter(
+      x => x.final_result.name === finalResult
+    ).length;
+    const percentage = (num / this.testResults.length) * 100;
     return percentage.toFixed(1) + '% | ' + num;
   }
 
   fillChartColors() {
-    const colors: any[] = [];
+    const backgroundColors: any[] = [];
     for (const finalResult of this.listOfFinalResults) {
       switch (finalResult.name) {
         case 'Passed':
-          colors.push('#009900');
+          backgroundColors.push(colors.success.fill);
           break;
         case 'Failed':
-          colors.push('#CC0000');
+          backgroundColors.push(colors.danger.fill);
           break;
         case 'In Progress':
-          colors.push('#FF6600');
+          backgroundColors.push(colors.warning.fill);
           break;
         case 'Not Executed':
-          colors.push('#3366FF');
+          backgroundColors.push(colors.primary.fill);
           break;
         case 'Pending':
-          colors.push('#5bc0de');
+          backgroundColors.push(colors.info.fill);
           break;
       }
     }
-    this.chartColors = [{ backgroundColor: colors }];
+    this.chartColors = [{ backgroundColor: backgroundColors }];
   }
 
   public chartClicked(event: any): void {
     if (event.active[0]) {
       const dataIndex = event.active[0]._index;
       const label: string = this.chart.labels[dataIndex].toString();
-      const clickedResult = this.listOfFinalResults.find(x => label.startsWith(x.name));
+      const clickedResult = this.listOfFinalResults.find(x =>
+        label.startsWith(x.name)
+      );
       this.clickedResult.emit(clickedResult);
     }
   }