+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 Features:
   - Move Administration to main Menu bar to make it more visible -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/58)
   - Test Run view: Add stability indicator -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/60)
+  - Add 'Show resolution slider' for Test Runs chart -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/43)
+
 
 Bugfixes:
   - Csv export with special symbols -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/36)