+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # CHANGELOG
 
-## 0.3.6 (unreleased)
+## 0.3.6 (2020-02-24)
 
 Features:
   - Move Administration to main Menu bar to make it more visible -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/58)
   - Test Run view: Add stability indicator -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/60)
   - Add 'Show resolution slider' for Test Runs chart -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/43)
-
+  - Move to Bootstrap 4.4.1 -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/66)
 
 Bugfixes:
   - Csv export with special symbols -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/36)